--- .eslintignore
@@ -1,3 +1,0 @@
-/lib
-/lib-esm
-/node_modules
