--- .eslintrc.yml
@@ -1,11 +1,0 @@
-extends:
-  - zakodium/ts
-  - zakodium/react
-  - zakodium/unicorn
-rules:
-  '@typescript-eslint/no-dynamic-delete': off
-  'import/namespace': off
-overrides:
-  - files: ['tests/**/*.test.tsx']
-    rules:
-      '@typescript-eslint/unbound-method': off
