--- .ncurc.json
@@ -1,3 +1,0 @@
-{
-  "reject": ["eslint"]
-}
