+++ .storybook/main.js
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   stories: ['../stories/**/*.stories.@(js|jsx|ts|tsx)'],
 
   addons: [
