+++ package.json
@@ -2,6 +2,7 @@
   "name": "react-plot",
   "version": "2.0.0",
   "description": "Library of React components to render SVG 2D plots.",
+  "type": "module",
   "main": "lib/index.js",
   "module": "lib-esm/index.js",
   "types": "lib-esm/index.d.ts",
@@ -14,6 +15,7 @@
     "build": "npm run clean && npm run build-ts",
     "build-storybook": "storybook build",
     "build-ts": "tsc -p tsconfig.cjs.json && tsc -p tsconfig.esm.json",
+    "postbuild-ts": "echo '{\"type\":\"commonjs\"}' > lib/package.json",
     "check-types": "tsc --noEmit",
     "clean": "rimraf lib lib-esm",
     "dev": "storybook dev -p 6006",
@@ -54,37 +56,38 @@
     "react": ">=18.0.0"
   },
   "devDependencies": {
-    "@playwright/experimental-ct-react": "^1.45.3",
+    "@playwright/experimental-ct-react": "^1.48.0",
     "@simbathesailor/use-what-changed": "^2.0.0",
-    "@storybook/addon-essentials": "^8.2.6",
-    "@storybook/addon-links": "^8.2.6",
-    "@storybook/addon-storysource": "^8.2.6",
-    "@storybook/react": "^8.2.6",
-    "@storybook/react-vite": "^8.2.6",
+    "@storybook/addon-essentials": "^8.3.5",
+    "@storybook/addon-links": "^8.3.5",
+    "@storybook/addon-storysource": "^8.3.5",
+    "@storybook/react": "^8.3.5",
+    "@storybook/react-vite": "^8.3.5",
     "@types/d3-array": "^3.2.1",
     "@types/d3-scale": "^4.0.8",
     "@types/d3-scale-chromatic": "^3.0.3",
     "@types/d3-shape": "^3.1.6",
-    "@types/node": "^22.0.0",
+    "@types/node": "^22.7.5",
     "@types/point-in-polygon": "^1.1.5",
-    "@types/react": "^18.3.3",
+    "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.0",
-    "eslint": "^8.57.0",
-    "eslint-config-zakodium": "^10.0.0",
-    "isotopic-distribution": "^3.2.1",
+    "eslint": "^9.12.0",
+    "eslint-config-zakodium": "^13.0.0",
+    "isotopic-distribution": "^3.3.2",
     "iv-analysis": "^0.3.0",
     "ml-dataset-iris": "^1.2.1",
     "ml-directional-distribution": "^0.1.1",
     "ml-pca": "^4.1.1",
     "ml-regression-simple-linear": "^3.0.1",
-    "ml-spectra-processing": "^14.5.1",
-    "ms-spectrum": "^3.6.2",
+    "ml-spectra-processing": "^14.5.3",
+    "ms-spectrum": "^3.6.7",
     "point-in-polygon": "^1.1.0",
     "prettier": "^3.3.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^4.0.13",
-    "typescript": "^5.5.4"
+    "rimraf": "^6.0.1",
+    "typescript": "^5.6.3"
   },
   "dependencies": {
     "d3-array": "^3.2.4",
@@ -96,6 +99,6 @@
     "react-d3-utils": "^1.0.0"
   },
   "volta": {
-    "node": "20.16.0"
+    "node": "22.9.0"
   }
 }
