+++ src/components/Annotations/Annotations.tsx
@@ -1,0 +1,9 @@
+import { type ReactElement, ReactNode } from 'react';
+
+export interface AnnotationsProps {
+  children: ReactNode;
+}
+
+export function Annotations(props: AnnotationsProps): ReactElement {
+  return <>{props.children}</>;
+}
