+++ src/components/Markers.map.ts
@@ -1,0 +1,23 @@
+import {
+  Circle,
+  Cross,
+  Diamond,
+  Hexagon,
+  Pentagon,
+  Square,
+  Star,
+  Triangle,
+  XMark,
+} from './Markers';
+
+export const markersMap = {
+  circle: Circle,
+  square: Square,
+  diamond: Diamond,
+  triangle: Triangle,
+  cross: Cross,
+  xmark: XMark,
+  pentagon: Pentagon,
+  star: Star,
+  hexagon: Hexagon,
+};
